Update Browsersync & Gulp and update gulpfile.js to Gulp 4 syntax
The Browsersync proxy URL http://localhost:8080 does not work with our
current Vagrant setup, but by instead using the development URL of the site
Browsersync works.

Gulp 4 requires changing the usage of gulp.series() in gulp.task() instead
of the old syntax gulp.task('example', ['someothertask'], ...

The usePolling parameter was also added to gulp.watch() because it was
found out that filesystem events are not synced between the host and VM as
described in https://github.com/floatdrop/gulp-watch/issues/213. This may
make gulp a tad more CPU-intensive, but nothing too drastic.

Closes: #107.